+++ src/com/google/gwt/maps/client/base/impl/PointImpl.java
@@ -25,16 +25,16 @@ import com.google.gwt.jsio.client.JSFlyweightWrapper;
 public interface PointImpl extends JSFlyweightWrapper {
 
   @Constructor("$wnd.google.maps.Point")
-  public JavaScriptObject construct(int x, int y);
+  public JavaScriptObject construct(double x, double y);
   
   public boolean equals(JavaScriptObject jso, JavaScriptObject other);
   
   public String toString(JavaScriptObject jso);
   
   @BeanProperties
-  public int getX(JavaScriptObject jso);
+  public double getX(JavaScriptObject jso);
   
   @BeanProperties
-  public int getY(JavaScriptObject jso);
+  public double getY(JavaScriptObject jso);
   
 }
